(chore): fix tests
