+++ packages/wdio-applitools-service/package.json
@@ -34,7 +34,7 @@
     "url": "https://github.com/webdriverio/webdriverio/issues"
   },
   "dependencies": {
-    "@applitools/eyes.webdriverio": "^1.1.0",
+    "@applitools/eyes.webdriverio": "^2.0.0",
     "wdio-logger": "^5.0.0-beta.3"
   }
 }
