+++ packages/wdio-config/tests/__mocks__/wdio-config.js
@@ -27,17 +27,16 @@ const pluginMocks = {
     }
 }
 
-export default {
-    initialisePlugin: jest.fn().mockImplementation((name, type) => pluginMocks[type][name]),
-    validateConfig: jest.fn().mockImplementation(
-        (_, config) => Object.assign(
-            DEFAULT_CONFIGS,
-            { hostname: '0.0.0.0', port: 4444, protocol: 'http', path: '/wd/hub' },
-            config
-        )
-    ),
-    wrapCommand: (_, origFn) => origFn,
-    ConfigParser: ConfigParserMock,
-    runTestInFiberContext: jest.fn(),
-    executeHooksWithArgs: jest.fn()
-}
+export const initialisePlugin = jest.fn().mockImplementation(
+    (name, type) => pluginMocks[type][name])
+export const validateConfig = jest.fn().mockImplementation(
+    (_, config) => Object.assign(
+        DEFAULT_CONFIGS,
+        { hostname: '0.0.0.0', port: 4444, protocol: 'http', path: '/wd/hub' },
+        config
+    )
+)
+export const wrapCommand = (_, origFn) => origFn
+export const ConfigParser = ConfigParserMock
+export const runTestInFiberContext = jest.fn()
+export const executeHooksWithArgs = jest.fn()
