+++ packages/webdriverio/tests/commands/element/saveScreenshot.test.js
@@ -39,7 +39,7 @@ describe('saveScreenshot', () => {
     })
 
     it('should fail if not existing directory', async () => {
-        const fs = require('fs')
+        const fs = require('fs').default
         fs.existsSync = () => false
 
         const browser = await remote({
